+++ src/main/java/model/dto/CreateChatConnectionUsingUserIDDto.java
@@ -4,7 +4,7 @@ import lombok.Builder;
 
 @Builder
 public record CreateChatConnectionUsingUserIDDto(
-        int loginUserId,
-        int userWantedToConnectId
+        Long loginUserId,
+        Long userWantedToConnectId
 ) {
 }
